Pull income documents into their own class
